tessi_list_tables

The list of Tessitura tables is configured in the extdata/tessi_tables.yml file in the package directory supplemented by the tessi_tables dictionary in config.yml

Usage

tessi_list_tables()

Value

data.table of configured Tessitura tables with columns short_name, long_name, base_table and primary_keys

Details

yml format

{short_name}:
   long_name: {name of table/view to be loaded}
   base_table: {the underlying table being queried that has primary_keys}
   primary_keys: {the primary key(s) as a value or a list of values}

Examples

# customers:
#   long_name: BI.VT_CUSTOMER
#   base_table: T_CUSTOMER
#   primary_keys: customer_no

tessi_list_tables()[short_name == "customers"]
#>    short_name      long_name base_table primary_keys date_column incremental
#>        <char>         <char>     <char>       <char>      <char>      <char>
#> 1:  customers BI.VT_CUSTOMER T_CUSTOMER  customer_no        <NA>        <NA>
#>     query
#>    <char>
#> 1:   <NA>